add some new files.

+++ src/sage/matrix/matrix_rational_dense.pyx
@@ -514,7 +514,7 @@ cdef class Matrix_rational_dense(matrix_dense.Matrix_dense):
 
     cdef int mpz_denom(self, mpz_t d) except -1:
         mpz_set_si(d,1)
-        cdef int i, j
+        cdef Py_ssize_t i, j
         cdef mpq_t *self_row
         _sig_on
         for i from 0 <= i < self._nrows:
